backend/requirements.txt
Flask==0.12.3
flask-cors===3.0.6
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-OpenID==1.2.5
SQLAlchemy==1.2.2
sqlalchemy-utils==0.32.21
psycopg2==2.7.3.2
tornado==4.5.3
pyjwt==1.5.3
steam==0.8.23
dota2==0.2.11
Flask-Limiter==1.0.1
websocket-client
grenouilleAPIClient==1.1.1
google-api-python-client==1.6.7
docker==3.4.1
Pillow==5.2.0
google-auth
google-auth-httplib2
Authlib
lru-dict
